move bb_task model tests to tasks app
